gad: attempt to always recalculate mask, currently broken
Always recalculates the mask. Unfortunately, this might bring permissions into
effect that weren't intended to be active and which aren't forced by gad.

Example: Executable bits for the ACL_GROUP entries are not filtered out by gad,
but might be suppressed by ACL_MASK. By recalculating ACL_MASK the executable
bit is always set to active. This needs to be fixed by recalculating the acl
entries according to the mask.