Minor refactoring, various additional comments
Moved most of the ACL code into a simpleacl module, as it's not specific to and
simply wraps around posix1e. Significantly cuts down on the size of gad.

While trying to understand the code again, I added a bunch of comments and
slightly moved around things. There should be no functional changes though.